lora: delete the output_dtype knob nothing in production sets
output_dtype was threaded through MoeLoraLayerEngine.run and
MoeLoraRunner.run_plan, and no production caller ever set it:
moe_runner/lora.py calls engine.run(dispatch_output, batch) with no kwarg,
so it always resolved to hidden_states.dtype. It is not the fp32 path the
DSA indexer LoRA needs either -- that is the dense LoRA output, in a
different code path, and nothing under lora/moe references the indexer.

Its only callers were seven test sites, and the reason they needed it turns
out to be smaller than it looked: torch.testing.assert_close checks dtype
by default, so a BF16 output compared against an FP32 CPU reference fails
on the dtype attribute before any tolerance is consulted. The parameter
existed to spare the tests four .float() calls.

Deleting it makes the numerics tests stricter, not looser. They now compare
the BF16 tensor production actually returns, against the same FP32
reference at the same tolerances (atol 0.018, rtol 0.06), where before they
compared an FP32 output no server ever produces. Both the eager and the
CUDA-graph replay comparisons hold unchanged, so BF16 rounding was never
what those tolerances were absorbing.

Verified on B200: 225 passed with the one pre-existing dense-gate failure.

+++ test/registered/unit/lora/test_moe_lora_contiguous_row_domain.py
@@ -811,9 +811,7 @@ def _run_once(
         is_prefill=is_prefill,
         has_active_lora=True,
     )
-    return runner.run_plan(
-        dispatch, batch, output_dtype=torch.float32, **runner._test_execution
-    )
+    return runner.run_plan(dispatch, batch, **runner._test_execution)
 
 
 # The one intended divergence: the masked and contiguous S2/S4 kernel
